Add files via upload
For height and weight about 22.2% and 23.2% respectively of the data is NA. For example we could replace them with the mean as you suggested of the most similar set of entries. For age its about 3.5% of the data given it's smaller we can dropping it when we're doing an analysis that requires looking at age is not a terrible choice.